Handle dbus events instead of signals
Emacs gets several signals related with different events of dbus which
makes it appear that it receives the same signals multiple times as commented in
Because we don't need to be listening all kinds of signals, we could
just listen for events over a dbus name by using `dbus-register-monitor`.

Along with this, the commit includes:
* Replace `darkman--mode-changed-signal-handler` with
`darkman--call-event-handler` to be a real handler of all method_call events
from darkman dbus.
* Create `darkman-load-theme-mode` function that will be called every
time `Mode` property in `nl.whynothugo.darkman` interface is called.
* Add Package-Requires Emacs 28.1 to fix warnings from package-lint
* Set related issue tasks as DONE
* Fix #10 as commented @grtcdr in #issuecomment-1447145236